Add FWGT chunk even if empty
Edited _write_clth to create the FWGT chunk even if the fixed weights aren't enveloped to any bones, this matches the ZETools exporter behavior as it appears to be a required chunk of CLTH chunks. All other child chunks are also assumed to be required except constraint chunks and COLL chunk.